demo/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ipware==4.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
